+++ creme/utils/__init__.py
@@ -6,6 +6,8 @@ from .math import norm
 from .math import prod
 from .math import sigmoid
 from .math import softmax
+from .window import Window
+from .window import SortedWindow
 
 
 __all__ = [
@@ -16,5 +18,7 @@ __all__ = [
     'norm',
     'prod',
     'sigmoid',
-    'softmax'
+    'softmax',
+    'SortedWindow',
+    'Window'
 ]